Moved AgentStarter into its own module to avoid a circular dependency.
